Introduce Jest to test test/utils.js
This increases the number of dependencies from 86 to 643.

Fixes https://github.com/mdn/browser-compat-data/issues/6855.
